DRY up mail notifier config validation and subject creation
